Add yarn 3+ workaround for missing mmdc file. (#240)
Yarn 3+ runs package bin scripts from the package.json `bin` section directly, instead of creating a file on the filesystem that points to the scripts listed. This means that there isn't an `mmdc` file created to point `prisma-erd-generator` to, so the user will need to create their own `mmdc` file. Added instructions on how to do that, with example script that can be used.